Update solver and example to use new dense and sparse headers; adjust matrix type handling
