+++ ezp/ezp.sparse
@@ -15,5 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
 
-#include "ezp.dense"
-#include "ezp.sparse"
+#include "lis.hpp"
+#include "mumps.hpp"
+#include "pardiso.hpp"
